Move autocomplete subscription into its own file
